requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==10.0.0
reportlab==3.6.12
pytest==7.2.2
pytest-django==4.5.2